requirements.txt
Flask==2.0.1
flask-cors
opencv-python-headless
Werkzeug==2.0.3
gunicorn
numpy
python-dotenv
Pillow
tensorflow
gdown
reportlab
fer
pandas
scikit-learn
scipy

